# docs/swagger.yaml
basePath: /
definitions:
...
  model.PairPhoneRequest:
    properties:
      phone:
        type: string
    type: object
...
  model.Response:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
  model.SessionConnectRequest:
    properties:
      immediate:
        example: true
        type: boolean
    type: object
  model.SessionCreateRequest:
    properties:
      events:
        type: string
      name:
        type: string
      proxyUrl:
        type: string
      webhook:
        type: string
    required:
    - name
    type: object
  model.SessionUpdateRequest:
    properties:
      events:
        type: string
      name:
        type: string
      proxyUrl:
        type: string
      webhook:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@fiozap.io
    name: FioZap Support
  description: WhatsApp API using whatsmeow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: FioZap API
  version: "1.0"
paths:
  /admin/sessions:
    get:
      description: List all WhatsApp sessions from all users (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - AdminKeyAuth: []
      summary: List all sessions (admin)
      tags:
      - Admin
...
  /sessions:
    get:
      description: List all WhatsApp sessions for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: List all sessions
      tags:
      - Sessions
    post:
      consumes:
      - application/json
      description: Create a new WhatsApp session
      parameters:
      - description: Session data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.SessionCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new session
      tags:
      - Sessions
  /sessions/{sessionId}:
    delete:
      description: Delete a WhatsApp session
      parameters:
      - description: Session name
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete session
      tags:
      - Sessions
    get:
      description: Get details of a specific WhatsApp session
      parameters:
      - description: Session name
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Get session details
      tags:
      - Sessions
    put:
      consumes:
      - application/json
      description: Update a WhatsApp session configuration
      parameters:
      - description: Session name
        in: path
        name: sessionId
        required: true
        type: string
      - description: Session data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.SessionUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Update session
      tags:
      - Sessions
  /sessions/{sessionId}/connect:
    post:
      consumes:
      - application/json
      description: Connect and start a WhatsApp session
      parameters:
      - description: Session name
        in: path
        name: sessionId
        required: true
        type: string
      - description: Connection options
        in: body
        name: request
        schema:
          $ref: '#/definitions/model.SessionConnectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Connect WhatsApp session
      tags:
      - Sessions
  /sessions/{sessionId}/disconnect:
    post:
      description: Disconnect the WhatsApp session
      parameters:
      - description: Session name
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Disconnect WhatsApp session
      tags:
      - Sessions
  /sessions/{sessionId}/logout:
    post:
      description: Logout and clear WhatsApp session data
      parameters:
      - description: Session name
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Logout WhatsApp session
      tags:
      - Sessions
  /sessions/{sessionId}/pairphone:
    post:
      consumes:
      - application/json
      description: Get pairing code for phone number authentication
      parameters:
      - description: Session name
        in: path
        name: sessionId
        required: true
        type: string
      - description: Phone number
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.PairPhoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Pair with phone number
      tags:
      - Sessions
  /sessions/{sessionId}/qr:
    get:
      description: Get QR code for WhatsApp authentication
      parameters:
      - description: Session name
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Get QR code
      tags:
      - Sessions
  /sessions/{sessionId}/status:
    get:
      description: Get the WhatsApp session status
      parameters:
      - description: Session name
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Get session status
      tags:
      - Sessions
...
securityDefinitions:
  AdminKeyAuth:
    description: Admin token for user management
    in: header
    name: Authorization
    type: apiKey
  ApiKeyAuth:
    description: User API token for session operations
    in: header
    name: Token
    type: apiKey
swagger: "2.0"